# public/20191114 cissukl.swagger.json
{
  "info": {
    "description": "Popis RESTful API",
    "version": "1.0.0",
    "title": "Číselníky SÚKL"
  },
  "host": "localhost:8000",
  "basePath": "/cissuklapi/v1",
  "consumes": [
    "application/json"
  ],
  "paths": {
...
    "/cenypuvodce": {
      "get": {
        "tags": [
          "Ceny puvodce"
        ],
        "description": "Vrací seznam cen původce",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fields",
            "in": "query",
            "description": "Kód SÚKL léčivého přípravku",
            "required": false,
            "type": "string",
            "enum": [
              "kod_sukl"
            ]
          },
          {
            "name": "kod_sukl",
            "in": "query",
            "description": "Kód SÚKL léčivého přípravku",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit, přednastavená hodnota limit=20",
            "required": false,
            "type": "number"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset, přednastavená hodnota offset=0",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/CenyPuvodce"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
    "/cenypuvodce/{kod_sukl}": {
      "get": {
        "tags": [
          "Ceny puvodce"
        ],
        "description": "Vrací cenu původce k danému kódu SÚKL",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "kod_sukl",
            "in": "path",
            "description": "Kód SÚKL",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "description": "Operace proběhla bez chyby",
            "schema": {
              "$ref": "#/definitions/CenyPuvodce"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
...
    "/lekarny": {
      "get": {
        "tags": [
          "Lekarny"
        ],
        "description": "Vrací seznam informací o lékárnách",
        "produces": [
          "text"
        ],
        "parameters": [
          {
            "name": "fields",
            "in": "query",
            "description": "Kód pracovistě",
            "required": false,
            "type": "string",
            "enum": [
              "kod_pracoviste"
            ]
          },
          {
            "name": "status",
            "in": "query",
            "description": "Status pracovistě",
            "required": false,
            "type": "string",
            "enum": [
              "zarazeno",
              "vyrazeno"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "schema": {
              "$ref": "#/definitions/Lekarny"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
    "/lekarny/{kod_pracoviste}": {
      "get": {
        "tags": [
          "Lekarny"
        ],
        "description": "Vrací informace lékárně pro daný kód pracoviště",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "kod_pracoviste",
            "in": "path",
            "description": "Kód pracoviste",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "schema": {
              "$ref": "#/definitions/Lekarny"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Error400": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Pro dané URL není služba implementována",
          "maxLength": 256
        }
      }
    },
    "Error404": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Nenalezeny žádné záznamy",
          "maxLength": 256
        }
      }
    },
...
    "CenyPuvodce": {
      "type": "object",
      "properties": {
        "KOD_SUKL": {
          "type": "string",
          "example": "0027193",
          "description": "Kód SÚKL",
          "maxLength": 7
        },
        "CENA_PUVODCE": {
          "type": "number",
          "example": "79623,21",
          "description": "Cena původce"
        }
      }
    },
...
    "Lekarny": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Lekarna"
      }
    },
    "Lekarna": {
      "type": "object",
      "properties": {
        "KOD_PRACOVISTE": {
          "type": "string",
          "example": "01995085000",
          "description": "Kód pracoviště",
          "maxLength": 11
        },
        "NAZEV": {
          "type": "string",
          "example": "Adamova lékárna",
          "description": "Název lékárny",
          "maxLength": 60
        },
        "KOD_LEKARNA": {
          "type": "string",
          "example": "01995680",
          "description": "Kód lékárny",
          "maxLength": 10
        },
        "KOD_TYP_LEKARNA": {
          "type": "string",
          "example": "Kód typu lékárny",
          "description": "Kód typu lékárny",
          "maxLength": 20
        },
        "ICO": {
          "type": "string",
          "example": "25136143",
          "description": "IČO",
          "maxLength": 12
        },
        "ICZ": {
          "type": "string",
          "example": "01995085",
          "description": "IČZ",
          "maxLength": 30
        },
        "MESTO": {
          "type": "string",
          "example": "Praha 1",
          "description": "Město lékárny",
          "maxLength": 50
        },
        "ULICE": {
          "type": "string",
          "example": "Václavské náměstí 8",
          "description": "Ulice lékárny",
          "maxLength": 30
        },
        "PSC": {
          "type": "string",
          "example": "110 10",
          "description": "PSČ",
          "maxLength": 10
        },
        "PROVOZOVATEL": {
          "type": "string",
          "example": "Pharma 97, s.r.o., Křovinovo náměstí 9/12, 193 00 Praha 9 - Horní Počernice",
          "description": "Provozovatel lékárny",
          "maxLength": 400
        },
        "LEKARNIK_PRIJMENI": {
          "type": "string",
          "example": "Fišerová",
          "description": "Příjmení lékárníka",
          "maxLength": 100
        },
        "LEKARNIK_JMENO": {
          "type": "string",
          "example": "Jana",
          "description": "Jméno lékárníka",
          "maxLength": 15
        },
        "LEKARNIK_TITUL": {
          "type": "string",
          "example": "Mgr.",
          "description": "Titul lékárníka",
          "maxLength": 20
        },
        "KOD_OKL": {
          "type": "string",
          "example": "401",
          "description": "Kód OKL",
          "maxLength": 5
        },
        "STATUS": {
          "type": "string",
          "example": "V",
          "description": "příznak statusu lékárny, nabývá hodnot \"Z\" - zařazena, \"V\" - vyřazena",
          "maxLength": 1
        },
        "REGION": {
          "type": "string",
          "example": "Praha  1",
          "description": "Region",
          "maxLength": 15
        },
        "WWW": {
          "type": "string",
          "example": "null",
          "description": "Webové adresy",
          "maxLength": 200
        },
        "EMAIL": {
          "type": "string",
          "example": "info@mantamedica.cz",
          "description": "Mailové adresy",
          "maxLength": 100
        },
        "TELEFON": {
          "type": "string",
          "example": "603 293 038",
          "description": "Telefon",
          "maxLength": 50
        },
        "FAX": {
          "type": "string",
          "example": "null",
          "description": "Fax",
          "maxLength": 100
        },
        "ERECEPT": {
          "type": "string",
          "example": "N",
          "description": "příznak, zda lékárna vydává erecpt,  nabývá hodnot \"A\" - ano, \"N\" - ne",
          "maxLength": 1
        },
        "POHOTOVOST_24H": {
          "type": "string",
          "example": "N",
          "description": "příznak, zda má lékárna pohotovost,  nabývá hodnot \"A\" - ano, \"N\" - ne",
          "maxLength": 1
        }
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "swagger": "2.0"
}